requirements.txt
ScrapyElasticSearch
